+++ Evergreen/Private/Get-Architecture.ps1
@@ -1,4 +1,4 @@
-Function Get-Architecture {
+function Get-Architecture {
     [OutputType([System.String])]
     [CmdletBinding(SupportsShouldProcess = $False)]
     param (
@@ -7,31 +7,33 @@ Function Get-Architecture {
         [System.String] $String
     )
 
-    Switch -Regex ($String.ToLower()) {
-        "aarch64" { $architecture = "ARM64"; Break }
-        "amd64" { $architecture = "AMD64"; Break }
-        "arm64" { $architecture = "ARM64"; Break }
-        "arm32" { $architecture = "ARM32"; Break }
-        "arm" { $architecture = "ARM32"; Break }
-        "win64" { $architecture = "x64"; Break }
-        "win32" { $architecture = "x86"; Break }
-        "win-arm64" { $architecture = "ARM64"; Break }
-        "win-x64" { $architecture = "x64"; Break }
-        "win-x86" { $architecture = "x86"; Break }
-        "x86_64" { $architecture = "x64"; Break }
-        "x64" { $architecture = "x64"; Break }
-        "w64" { $architecture = "x64"; Break }
-        "-64" { $architecture = "x64"; Break }
-        "64-bit" { $architecture = "x64"; Break }
-        "64bit" { $architecture = "x64"; Break }
-        "32-bit" { $architecture = "x86"; Break }
-        "32bit" { $architecture = "x86"; Break }
-        "x32" { $architecture = "x86"; Break }
-        "w32" { $architecture = "x86"; Break }
-        "-32" { $architecture = "x86"; Break }
-        "-x86" { $architecture = "x86"; Break }
-        "x86" { $architecture = "x86"; Break }
-        Default {
+    switch -Regex ($String.ToLower()) {
+        "aarch64" { $architecture = "ARM64"; break }
+        "amd64" { $architecture = "AMD64"; break }
+        "arm64" { $architecture = "ARM64"; break }
+        "arm32" { $architecture = "ARM32"; break }
+        "arm" { $architecture = "ARM32"; break }
+        "nt64" { $architecture = "x64"; break }
+        "nt32" { $architecture = "x86"; break }
+        "win64" { $architecture = "x64"; break }
+        "win32" { $architecture = "x86"; break }
+        "win-arm64" { $architecture = "ARM64"; break }
+        "win-x64" { $architecture = "x64"; break }
+        "win-x86" { $architecture = "x86"; break }
+        "x86_64" { $architecture = "x64"; break }
+        "x64" { $architecture = "x64"; break }
+        "w64" { $architecture = "x64"; break }
+        "-64" { $architecture = "x64"; break }
+        "64-bit" { $architecture = "x64"; break }
+        "64bit" { $architecture = "x64"; break }
+        "32-bit" { $architecture = "x86"; break }
+        "32bit" { $architecture = "x86"; break }
+        "x32" { $architecture = "x86"; break }
+        "w32" { $architecture = "x86"; break }
+        "-32" { $architecture = "x86"; break }
+        "-x86" { $architecture = "x86"; break }
+        "x86" { $architecture = "x86"; break }
+        default {
             Write-Verbose -Message "$($MyInvocation.MyCommand): Architecture not found in $String, defaulting to x86."
             $architecture = "x86"
         }